ollama/tutorials/requirements.txt
langchain==0.3.7
langchain-ollama==0.2.0
python-dotenv==1.0.1